Replace @line field with @range in Token and Node.
Previously the @line field that some Nodes had wasn't always
the number of the line where a Node started.  And it was hard
to tell why some nodes had this info, while others didn't.

Now both the start and end line of every Token and Node is
stored in @range field as two-element array.

The #line method is still there for backwards compatibility,
but the @range field is the one that holds the actual data.

For now I had to disable the tests in test_comments.rb as
the #apply_comments (which it effectively tested) relied on
only some nodes having the line number data available.