Map lexemes to possible start-characters.
Instead of looping through the array of all lexemes on every step,
create a map of characters-->lexemes telling which lexemes can
begin with a certain character.

So that when tokenizing we peek the next character, look up those
few lexemes that can begin with it, and only try to mach these.

My benchmarks show 2 x tokenizing performance inrease with this
optimization.