Reorder lexemes to only look for the first match.
By just moving the :LITERALS lexeme between :REGEXP and :SINGLE_CHAR,
the order of lexemes is now such that we can now blindly return the
first one that matches.

Also moved the :S lexeme alongside other one-line definitions.

Because of the previous char-lookup-table optimization, this one
improves the speed only so little.  But IMHO the code is a bit cleaner
this way.